Implement dry-run

+++ pkg/github/github.go
@@ -129,7 +129,11 @@ func NewClient(token string) *Client {
 	}
 }
 
-func (c *Client) SyncLabels(ctx context.Context, owner, repo string, labels []Label, prune bool) error {
+func (c *Client) SyncLabels(ctx context.Context, owner, repo string, labels []Label, prune bool, dryRun bool) error {
+	if dryRun {
+		fmt.Printf("Dry run! No actual changes will be made.\n")
+	}
+
 	labelMap := make(map[string]Label)
 	for _, l := range labels {
 		labelMap[l.Name] = l
@@ -155,7 +159,7 @@ func (c *Client) SyncLabels(ctx context.Context, owner, repo string, labels []La
 				if ok {
 					return nil
 				}
-				return c.deleteLabel(ctx, owner, repo, currentLabel.Name)
+				return c.deleteLabel(ctx, owner, repo, currentLabel.Name, dryRun)
 			})
 		}
 
@@ -170,10 +174,10 @@ func (c *Client) SyncLabels(ctx context.Context, owner, repo string, labels []La
 		eg.Go(func() error {
 			currentLabel, ok := currentLabelMap[l.Name]
 			if !ok {
-				return c.createLabel(ctx, owner, repo, l)
+				return c.createLabel(ctx, owner, repo, l, dryRun)
 			}
 			if currentLabel.Description != l.Description || currentLabel.Color != l.Color {
-				return c.updateLabel(ctx, owner, repo, l)
+				return c.updateLabel(ctx, owner, repo, l, dryRun)
 			}
 			fmt.Printf("label: %+v not changed on %s/%s\n", l, owner, repo)
 			return nil
@@ -183,14 +187,17 @@ func (c *Client) SyncLabels(ctx context.Context, owner, repo string, labels []La
 	return eg.Wait()
 }
 
-func (c *Client) createLabel(ctx context.Context, owner, repo string, label Label) error {
+func (c *Client) createLabel(ctx context.Context, owner, repo string, label Label, dryRun bool) error {
 	l := &github.Label{
 		Name:        &label.Name,
 		Description: &label.Description,
 		Color:       &label.Color,
 	}
+	fmt.Printf("label create: %+v on: %s/%s\n", label, owner, repo)
+	if dryRun {
+		return nil
+	}
 	_, _, err := c.githubClient.Issues.CreateLabel(ctx, owner, repo, l)
-	fmt.Printf("label: %+v created on: %s/%s\n", label, owner, repo)
 	return err
 }
 
@@ -219,19 +226,25 @@ func (c *Client) getLabels(ctx context.Context, owner, repo string) ([]Label, er
 	return labels, nil
 }
 
-func (c *Client) updateLabel(ctx context.Context, owner, repo string, label Label) error {
+func (c *Client) updateLabel(ctx context.Context, owner, repo string, label Label, dryRun bool) error {
 	l := &github.Label{
 		Name:        &label.Name,
 		Description: &label.Description,
 		Color:       &label.Color,
 	}
+	fmt.Printf("label update %+v on: %s/%s\n", label, owner, repo)
+	if dryRun {
+		return nil
+	}
 	_, _, err := c.githubClient.Issues.EditLabel(ctx, owner, repo, label.Name, l)
-	fmt.Printf("label %+v updated on: %s/%s\n", label, owner, repo)
 	return err
 }
 
-func (c *Client) deleteLabel(ctx context.Context, owner, repo, name string) error {
+func (c *Client) deleteLabel(ctx context.Context, owner, repo, name string, dryRun bool) error {
+	fmt.Printf("label delete: %s from: %s/%s\n", name, owner, repo)
+	if dryRun {
+		return nil
+	}
 	_, err := c.githubClient.Issues.DeleteLabel(ctx, owner, repo, name)
-	fmt.Printf("label: %s deleted from: %s/%s\n", name, owner, repo)
 	return err
 }
